pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "google-generativeai>=0.8.5",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "streamlit>=1.44.1",
]
